pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "loreforge"
version = "0.1.0"
description = "Azure data pipeline for tabletop RPG data"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "psycopg[binary]>=3.1",
    "structlog>=24.0",
    "python-ulid>=2.0",
    "pyarrow>=14.0",
    "azure-storage-blob>=12.0",
]

[project.scripts]
loreforge = "worker:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
mypy_path = "src"
